requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cattrs==24.1.3
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.10
cron-descriptor==1.4.5
Django==5.1.8
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-extensions==4.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
idna==3.10
kombu==5.5.2
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
svgwrite==1.4.3
Tree==0.2.4
tzdata==2025.2
url-normalize==2.2.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
